Refactor sample app activity
